requirements.txt
duckduckgo-search
streamlit
langchain_groq
langchain
langchain_community
python-dotenv